test(xychart): organize test/ into subfolders that mirror src/
