+++ packages/visx-xychart/test/enhancers/withRegisteredData.test.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react';
 import { mount } from 'enzyme';
-import withRegisteredData from '../src/enhancers/withRegisteredData';
-import getDataContext from './mocks/getDataContext';
-import { DataContext } from '../src';
+import withRegisteredData from '../../src/enhancers/withRegisteredData';
+import getDataContext from '../mocks/getDataContext';
+import { DataContext } from '../../src';
 
 const series = {
   key: 'visx',
