+++ packages/visx-xychart/test/providers/DataProvider.test.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { mount } from 'enzyme';
-import { DataProvider, DataContext } from '../src';
-import { DataProviderProps } from '../lib/providers/DataProvider';
+import { DataProvider, DataContext } from '../../src';
+import { DataProviderProps } from '../../lib/providers/DataProvider';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const getWrapper = (consumer: React.ReactNode, props?: DataProviderProps<any, any>) => {
